+++ src/main/java/com/prgrms/needit/domain/user/user/controller/UserController.java
@@ -6,12 +6,8 @@ import com.prgrms.needit.domain.user.user.dto.Response;
 import com.prgrms.needit.domain.user.user.service.UserService;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -22,28 +18,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
 	private final UserService userService;
-
-	@PostMapping("/login")
-	public ResponseEntity<ApiResponse<Response.TokenInfo>> login(
-		@Valid @RequestBody Request.Login login
-	) {
-		return ResponseEntity.ok(ApiResponse.of(userService.login(login)));
-	}
-
-	@PutMapping("/reissue")
-	public ResponseEntity<ApiResponse<Response.TokenInfo>> reissue(
-	@Valid @RequestBody Request.Reissue reissue
-	) {
-		return ResponseEntity.ok(ApiResponse.of(userService.reissue(reissue)));
-	}
-
-	@DeleteMapping("/logout")
-	public ResponseEntity<Void> logout(
-		@Valid @RequestBody Request.Logout logout
-	) {
-		userService.logout(logout);
-		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-	}
 
 	@PostMapping("/check-email")
 	public ResponseEntity<ApiResponse<Response.IsUnique>> checkEmail(
@@ -61,11 +35,6 @@ public class UserController {
 		return ResponseEntity.ok(
 			ApiResponse.of(userService.isNicknameUnique(request))
 		);
-	}
-
-	@GetMapping
-	public ResponseEntity<ApiResponse<Response.UserInfo>> getCurUser() {
-		return ResponseEntity.ok(ApiResponse.of(userService.getUserInfo()));
 	}
 
 }
