+++ src/main/java/com/prgrms/needit/domain/user/user/repository/UserRepository.java
@@ -7,4 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface UserRepository extends JpaRepository<Users, Long> {
 
 	Optional<Users> findByEmailAndIsDeletedFalse(String email);
+
+	boolean existsByEmail(String checkEmail);
+
+	boolean existsByNickname(String nickname);
 }
