+++ src/main/java/com/prgrms/needit/domain/user/user/service/CustomUserDetailsService.java
@@ -2,12 +2,10 @@ package com.prgrms.needit.domain.user.user.service;
 
 import com.prgrms.needit.common.error.ErrorCode;
 import com.prgrms.needit.common.error.exception.NotFoundResourceException;
-import com.prgrms.needit.domain.user.center.entity.Center;
-import com.prgrms.needit.domain.user.center.repository.CenterRepository;
-import com.prgrms.needit.domain.user.member.entity.Member;
-import com.prgrms.needit.domain.user.member.repository.MemberRepository;
+import com.prgrms.needit.domain.user.user.entity.Users;
+import com.prgrms.needit.domain.user.user.repository.UserRepository;
 import java.util.Collections;
-import java.util.Optional;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -16,43 +14,25 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class CustomUserDetailsService implements UserDetailsService {
 
-	private final MemberRepository memberRepository;
-	private final CenterRepository centerRepository;
-
-	public CustomUserDetailsService(
-		MemberRepository memberRepository,
-		CenterRepository centerRepository
-	) {
-		this.memberRepository = memberRepository;
-		this.centerRepository = centerRepository;
-	}
+	private final UserRepository userRepository;
 
 	@Override
 	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-		Optional<Member> member = memberRepository.findByEmailAndIsDeletedFalse(email);
-		if (member.isPresent()) {
-			return createMemberDetails(member.get());
-		}
 
-		Optional<Center> center = centerRepository.findByEmailAndIsDeletedFalse(email);
-		if (center.isPresent()) {
-			return createCenterDetails(center.get());
-		}
-
-		throw new NotFoundResourceException(ErrorCode.NOT_FOUND_USER);
+		return userRepository.findByEmailAndIsDeletedFalse(email)
+							 .map(this::createUserDetails)
+							 .orElseThrow(() ->
+											  new NotFoundResourceException(
+												  ErrorCode.NOT_FOUND_USER));
 	}
 
-	private UserDetails createMemberDetails(Member member) {
-		return new User(member.getEmail(), member.getPassword(), Collections.singleton(
-			new SimpleGrantedAuthority(member.getUserRole()
-											 .getKey())));
+	private UserDetails createUserDetails(Users user) {
+		return new User(user.getEmail(), user.getPassword(), Collections.singleton(
+			new SimpleGrantedAuthority(user.getUserRole()
+										   .getKey())));
 	}
 
-	private UserDetails createCenterDetails(Center center) {
-		return new User(center.getEmail(), center.getPassword(), Collections.singleton(
-			new SimpleGrantedAuthority(center.getUserRole()
-											 .getKey())));
-	}
 }
